third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGRadialGradientElement-dom-fx-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'fx' attribute of the SVGRadialGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS radialGradientElement.getAttribute('fx') is "0%"
PASS radialGradientElement.getAttribute('fx') is "50%"
PASS successfullyParsed is true

TEST COMPLETE
